migrate to Nuxt
