--- .eslintrc.js
@@ -1,14 +1,0 @@
-module.exports = {
-  root: true,
-  env: {
-    node: true,
-  },
-  extends: [
-    'plugin:vue/vue3-essential',
-    'eslint:recommended',
-    '@vue/eslint-config-typescript',
-  ],
-  rules: {
-    'vue/multi-word-component-names': 'off',
-  },
-}
