+++ .github/workflows/build.yml
@@ -20,10 +20,10 @@ jobs:
         run: npm ci
 
       - name: Compile
-        run: npm run build
+        run: npm run generate
 
       - name: Deploy
         uses: peaceiris/actions-gh-pages@v3
         with:
           deploy_key: ${{ secrets.ACTIONS_DEPLOY_KEY }}
-          publish_dir: ./dist
+          publish_dir: ./.output/public
