+++ app.vue
@@ -1,0 +1,5 @@
+<template>
+  <div>
+    <NuxtPage style="min-height: 100vh" />
+  </div>
+</template>
