+++ package.json
@@ -1,30 +1,19 @@
 {
-  "name": "localsend-website",
-  "version": "0.0.0",
+  "private": true,
   "scripts": {
-    "dev": "vite",
-    "build": "vue-tsc --noEmit && vite build",
-    "preview": "vite preview",
-    "lint": "eslint . --fix --ignore-path .gitignore"
+    "build": "nuxt build",
+    "dev": "nuxt dev",
+    "generate": "nuxt generate",
+    "preview": "nuxt preview",
+    "postinstall": "nuxt prepare"
   },
   "dependencies": {
-    "roboto-fontface": "*",
-    "vue": "^3.2.38",
-    "vue-router": "^4.0.0",
-    "vuetify": "^3.0.0",
-    "webfontloader": "^1.0.0"
+    "sass": "^1.60.0",
+    "vuetify": "^3.1.11"
   },
   "devDependencies": {
     "@mdi/js": "^7.2.96",
-    "@types/node": "^18.11.9",
-    "@types/webfontloader": "^1.6.35",
-    "@vitejs/plugin-vue": "^3.0.3",
-    "@vue/eslint-config-typescript": "^11.0.0",
-    "eslint": "^8.22.0",
-    "eslint-plugin-vue": "^9.3.0",
-    "typescript": "^4.0.0",
-    "vite": "^3.0.9",
-    "vite-plugin-vuetify": "^1.0.0-alpha.12",
-    "vue-tsc": "^1.0.9"
+    "@nuxtjs/google-fonts": "^3.0.0",
+    "nuxt": "^3.3.1"
   }
 }
