+++ plugins/vuetify.ts
@@ -1,0 +1,33 @@
+// plugins/vuetify.js
+import { defineNuxtPlugin } from "#imports";
+import { createVuetify } from 'vuetify'
+import * as components from 'vuetify/components'
+import * as directives from 'vuetify/directives'
+import { aliases, mdi } from 'vuetify/iconsets/mdi-svg'
+
+export default defineNuxtPlugin(nuxtApp => {
+    const vuetify = createVuetify({
+        ssr: true,
+        components,
+        directives,
+        theme: {
+            themes: {
+                light: {
+                    colors: {
+                        primary: '#009688',
+                        secondary: '#5CBBF6',
+                    },
+                },
+            },
+        },
+        icons: {
+            defaultSet: 'mdi',
+            aliases,
+            sets: {
+                mdi,
+            },
+        },
+    });
+
+    nuxtApp.vueApp.use(vuetify)
+});
