+++ public/img/badges/f-droid-badge.png
