--- src/views/Contact.vue
@@ -1,9 +1,0 @@
-<template>
-  <legal-layout>
-    <h1 class="text-h6">Contact: localsendapp@gmail.com</h1>
-  </legal-layout>
-</template>
-
-<script setup lang="ts">
-  import LegalLayout from "@/layouts/LegalLayout.vue";
-</script>
